+++ include/server/game/object.hpp
@@ -13,21 +13,6 @@ class SharedObject;
 struct SharedPhysics;
 
 /**
- * @brief An enum for the type of an object; the fields here should match all
- * class names in the inheritance tree in which Object is the root.
- */
-enum class ObjectType {
-	Object	//	Generic object type (base class)
-};
-
-/**
- * @brief Returns a string representation of the ObjectType enum
- * @param type ObjectType enum
- * @return String representation of the given ObjectType enum
- */
-std::string objectTypeString(ObjectType type);
-
-/**
  * @brief Physics struct that contains all movement / collision related data for
  * a particular object
  */
