+++ include/server/game/servergamestate.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "shared/utilities/typedefs.hpp"
+#include "shared/game/sharedgamestate.hpp"
 #include "shared/utilities/config.hpp"
 #include "shared/utilities/smartvector.hpp"
 #include "server/game/object.hpp"
@@ -13,34 +14,6 @@
 
 //	From sharedgamestate.hpp
 struct SharedGameState;
-
-enum class GamePhase {
-	TITLE_SCREEN,
-	LOBBY,
-	GAME
-};
-
-/**
- * @brief Information about the current lobby of players.
- */
-struct Lobby {
-	//	TODO: Perhaps instead of a mapping from EntityID -> string, the mapping
-	//	could eventually be EntityID -> Player (where Player derives from
-	//	Object)?
-
-	/**
-	 * @brief A hash table that maps from player's EntityID to their names.
-	 */
-	std::unordered_map<EntityID, std::string> players;
-
-	/**
-	 * @brief The maximum number of players that this game instance can support.
-	 */
-	int max_players;
-
-	//	TODO: Add a player role listing? I.e., which player is playing which
-	//	character and which player is playing as the Dungeon Master?
-};
 
 /*	Constants	*/
 #define	FIRST_TIMESTEP			0
